feat(auth): implement registration and logout actions; refactor signup page to use new action
feat(events): create EventsContent component for event listing and filtering; update EventsPage to fetch events

feat(event): add EventRegistrationButton component for event registration; update EventDetails to use new button

refactor(authService): remove authService and integrate eventService for event-related API calls
