+++ frontend/app/actions/auth.ts
@@ -1,10 +1,34 @@
 "use server";
 
 import { cookies } from "next/headers";
-import { SignInData, AuthResponse } from "@/types/auth";
+import { SignInData, AuthResponse, SignUpData } from "@/types/auth";
 import { redirect } from "next/navigation";
 
 const API_URL = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8080";
+
+export async function registerAction(formData: SignUpData) {
+  try {
+    const response = await fetch(`${API_URL}/auth/register`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(formData),
+    });
+
+    if (!response.ok) {
+      const errorData = await response.json().catch(() => ({}));
+      return {
+        error: errorData.message || "Erro ao criar conta. Tente novamente.",
+      };
+    }
+  } catch (error) {
+    console.error("Erro no cadastro:", error);
+    return { error: "Servidor indisponível. Tente novamente mais tarde." };
+  }
+
+  redirect("/signin");
+}
 
 export async function loginAction(formData: SignInData) {
   let data: AuthResponse;
@@ -41,6 +65,25 @@ export async function loginAction(formData: SignInData) {
 
 export async function logoutAction() {
   const cookieStore = await cookies();
+
+  const token = cookieStore.get("token")?.value;
+
+  if (token) {
+    try {
+      await fetch(`${API_URL}/auth/logout`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+      });
+      console.log("Logout realizado no backend com sucesso.");
+    } catch (error) {
+      console.error("Erro ao realizar logout no backend:", error);
+    }
+  }
+
   cookieStore.delete("token");
+
   redirect("/signin");
 }
