--- frontend/services/authService.ts
@@ -1,55 +1,0 @@
-import { SignInData, AuthResponse, SignUpData } from "@/types/auth";
-
-const API_URL = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8080";
-
-export const authService = {
-  login: async (credentials: SignInData): Promise<AuthResponse> => {
-    const response = await fetch(`${API_URL}/auth/login`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(credentials),
-    });
-
-    if (!response.ok) {
-      const data = await response.json().catch(() => ({}));
-      throw new Error(
-        data.message || "Credenciais inválidas. Verifique seu e-mail e senha.",
-      );
-    }
-
-    return response.json();
-  },
-
-  register: async (data: SignUpData): Promise<void> => {
-    const response = await fetch(`${API_URL}/auth/register`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (!response.ok) {
-      const errorData = await response.json().catch(() => ({}));
-      throw new Error(
-        errorData.message || "Erro ao criar conta. Tente novamente.",
-      );
-    }
-  },
-
-  logout: async (token: string): Promise<void> => {
-    try {
-      await fetch(`${API_URL}/auth/logout`, {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-      });
-    } catch (error) {
-      console.error("Erro silencioso no logout:", error);
-    }
-  },
-};
